requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.7.1
torchvision==0.22.1
Pillow>=11.3.0
matplotlib>=3.10.3
numpy>=2.3.1
pandas>=2.3.0
pytorch-metric-learning==2.8.1
scikit-learn>=1.7.0
timm==1.0.16
torch-ema>=0.3.0
torchmetrics==1.7.3
tqdm>=4.67.1
ttach==0.0.3
